# api/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /indexer/tx/v1/txs:
    get:
      consumes:
      - application/json
      description: Get a list of transactions with pagination
      parameters:
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: Get transactions
      tags:
      - Transactions
  /indexer/tx/v1/txs/{tx_hash}:
    get:
      consumes:
      - application/json
      description: Get a specific transaction by its hash
      parameters:
      - description: Transaction hash
        in: path
        name: tx_hash
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get transaction by hash
      tags:
      - Transactions
  /indexer/tx/v1/txs/by_account/{account}:
    get:
      consumes:
      - application/json
      description: Get transactions associated with a specific account
      parameters:
      - description: Account address
        in: path
        name: account
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: Get transactions by account
      tags:
      - Transactions
  /indexer/tx/v1/txs/by_height/{height}:
    get:
      consumes:
      - application/json
      description: Get transactions at a specific block height
      parameters:
      - description: Block height
        in: path
        name: height
        required: true
        type: integer
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: Get transactions by height
      tags:
      - Transactions
...
swagger: "2.0"